@@ -3,8 +3,8 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
+    "serve": "vite",
+    "build": "vite build",
     "test:unit": "vue-cli-service test:unit",
     "lint": "vue-cli-service lint"
   },
@@ -12,27 +12,32 @@
     "node": ">= 12.x"
   },
   "dependencies": {
-    "buefy": "0.9.8",
+    "@vue/compiler-sfc": "^3.0.5",
     "lodash": "4.17.21",
-    "vue": "2.6.14"
+    "vue": "^3.0.5"
   },
   "devDependencies": {
+    "@vitejs/plugin-vue": "^1.1.4",
     "@vue/cli-plugin-babel": "4.5.13",
     "@vue/cli-plugin-eslint": "4.5.13",
     "@vue/cli-plugin-unit-jest": "4.5.13",
     "@vue/cli-service": "4.5.13",
     "@vue/eslint-config-prettier": "6.0.0",
     "@vue/eslint-config-standard": "6.0.0",
-    "@vue/test-utils": "1.2.1",
+    "@vue/test-utils": "^2.0.0-rc.9",
+    "autoprefixer": "^10.2.4",
     "axios": "0.21.1",
     "babel-eslint": "10.1.0",
     "eslint": "7.29.0",
     "eslint-plugin-prettier": "3.4.0",
     "eslint-plugin-vue": "7.12.1",
     "fibers": "5.0.0",
+    "postcss": "^8.2.6",
     "prettier": "2.3.1",
     "sass": "1.35.1",
     "sass-loader": "12.1.0",
-    "vue-template-compiler": "2.6.14"
+    "tailwindcss": "^2.0.3",
+    "vite": "^2.0.1",
+    "vue-jest": "^5.0.0-alpha.10"
   }
 }
