@@ -5,7 +5,7 @@ module.exports = {
     node: true,
   },
 
-  extends: ['plugin:vue/essential', '@vue/prettier'],
+  extends: ['plugin:vue/vue3-essential', '@vue/prettier'],
 
   rules: {
     'no-console': process.env.NODE_ENV === 'production' ? 'error' : 'off',
