@@ -1,7 +1,9 @@
 module.exports = {
   preset: '@vue/cli-plugin-unit-jest',
-  setupFiles: ['./tests/setup.js'],
   collectCoverage: true,
   collectCoverageFrom: ['src/**/*.{js,vue}', '!src/main.js', '!src/App.vue'],
   coverageReporters: ['html', 'text'],
+  transform: {
+    '^.+\\.vue$': 'vue-jest',
+  },
 };
