@@ -17,7 +17,7 @@ describe('SearchResultList.vue', () => {
       describe('click button', () => {
         beforeEach(() => {
           wrapper = shallowMount(SearchResultList, {
-            propsData: {
+            props: {
               results: [result],
               nextUrl: 'next-url',
               isLoading: false,
@@ -40,7 +40,7 @@ describe('SearchResultList.vue', () => {
 
           beforeEach(() => {
             wrapper = shallowMount(SearchResultList, {
-              propsData: {
+              props: {
                 results: [result],
                 nextUrl: 'next-url',
                 isLoading: true,
@@ -49,12 +49,8 @@ describe('SearchResultList.vue', () => {
             button = wrapper.find('#more-results-btn');
           });
 
-          it('there is class is-loading of button', () => {
-            expect(button.classes()).toContain('is-loading');
-          });
-
           it('disable button', () => {
-            expect(button.attributes().disabled).toBe('disabled');
+            expect(button.attributes().disabled).toBeDefined();
           });
         });
 
@@ -63,17 +59,13 @@ describe('SearchResultList.vue', () => {
 
           beforeEach(async () => {
             wrapper = shallowMount(SearchResultList, {
-              propsData: {
+              props: {
                 results: [result],
                 nextUrl: 'next-url',
                 isLoading: false,
               },
             });
             button = wrapper.find('#more-results-btn');
-          });
-
-          it('there is not class is-loading of button', () => {
-            expect(button.classes()).not.toContain('is-loading');
           });
 
           it('enable button', () => {
@@ -87,7 +79,7 @@ describe('SearchResultList.vue', () => {
       describe('There are no results', () => {
         beforeEach(() => {
           wrapper = shallowMount(SearchResultList, {
-            propsData: {
+            props: {
               results: [],
               isNotFound: true,
             },
@@ -102,7 +94,7 @@ describe('SearchResultList.vue', () => {
       describe('There are results', () => {
         beforeEach(() => {
           wrapper = shallowMount(SearchResultList, {
-            propsData: {
+            props: {
               results: [],
               isNotFound: false,
             },
@@ -119,7 +111,7 @@ describe('SearchResultList.vue', () => {
       describe('There is an error', () => {
         beforeEach(() => {
           wrapper = shallowMount(SearchResultList, {
-            propsData: {
+            props: {
               results: [],
               isError: true,
             },
@@ -134,7 +126,7 @@ describe('SearchResultList.vue', () => {
       describe('There are no errors', () => {
         beforeEach(() => {
           wrapper = shallowMount(SearchResultList, {
-            propsData: {
+            props: {
               results: [],
               isError: false,
             },
@@ -159,7 +151,7 @@ describe('SearchResultList.vue', () => {
 
       it('Return true if there is nextUrl', () => {
         wrapper = shallowMount(SearchResultList, {
-          propsData: {
+          props: {
             results: [result, result],
             nextUrl: 'next-url',
           },
@@ -170,7 +162,7 @@ describe('SearchResultList.vue', () => {
 
       it('Return false if there is not nextUrl', () => {
         wrapper = shallowMount(SearchResultList, {
-          propsData: {
+          props: {
             results: [result, result],
             nextUrl: undefined,
           },
